+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support new `compose.yaml` default name.
 - Add option `-p, --profile` to select compose profiles.
 - Add option `-a, --all` to select all compose profiles.
+- Add option `-t, --timestamps` to show timestamps in logs.
 
 ### Changed
 - Verbose option changed from `-a, --all` to `-V, --verbose`.